requirements.txt
numpy~=1.26.4
opencv-python~=4.8.0.76
matplotlib~=3.8.4
pandas~=2.2.1
torch~=2.2.2
torchvision~=0.17.2
utils~=1.0.2
torchmetrics~=1.4.0.post0
fvcore~=0.1.5.post20221221
omegaconf~=2.3.0
setuptools~=68.2.2
pillow~=10.4.0
imageio~=2.34.0
scikit-image~=0.22.0
websockets~=12.0
kornia~=0.6.9
h5py~=3.10.0
scipy~=1.12.0
defusedxml~=0.7.1
ipython~=8.18.1
future~=1.0.0
pip~=23.2.1
emoji~=2.12.1
attrs~=25.1.0
colorlog~=6.8.2
toml~=0.10.2
wheel~=0.41.3
Brotli~=1.1.0
contrib~=0.3.0
cryptography~=41.0.7
lxml~=5.2.1
tornado~=6.4.1
Jinja2~=3.1.3
common~=0.1.2
pytest~=8.3.4
pytz~=2024.1
cffi~=1.17.1
numba~=0.59.1
sympy~=1.13.3
mpmath~=1.3.0
symengine~=0.11.0
decorator~=4.4.2
tabulate~=0.9.0
pydot~=2.0.0
astunparse~=1.6.3
onnx~=1.16.1
tensorboard~=2.9.0
filelock~=3.13.3
tqdm~=4.66.5
tensorflow~=2.9.0
psutil~=5.9.8
colorama~=0.4.6
redis~=5.0.4
autograd~=1.8.0
networkx~=2.8.8
urllib3~=1.26.20
fsspec~=2024.3.1
requests~=2.32.3
aiohttp~=3.9.5
yarl~=1.9.4
MarkupSafe~=2.1.5
PyYAML~=6.0.2
soundfile~=0.12.1
ops~=2.17.1
efficientnet~=1.1.1
seaborn~=0.13.2
scikit-learn~=1.5.2
termcolor~=1.1.0
customtkinter~=5.2.2
keras~=2.9.0
lightgbm~=4.3.0
transformers~=4.46.3